Refactor CoastGuard; move from executor to monitor
